More refactoring. Starting unit tests for streaming SSR.
