+++ src/Components/Endpoints/src/RazorComponentEndpoint.cs
@@ -10,7 +10,6 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.WebUtilities;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Hosting.Internal;
 using Microsoft.Extensions.Options;
 
 namespace Microsoft.AspNetCore.Components.Endpoints;
@@ -20,14 +19,15 @@ internal static class RazorComponentEndpoint
     public static RequestDelegate CreateRouteDelegate(Type componentType)
     {
         return httpContext =>
-            RenderComponentToResponse(httpContext, RenderMode.Static, componentType, componentParameters: null);
+            RenderComponentToResponse(httpContext, RenderMode.Static, componentType, componentParameters: null, preventStreamingRendering: false);
     }
 
     internal static Task RenderComponentToResponse(
         HttpContext httpContext,
         RenderMode renderMode,
         Type componentType,
-        IReadOnlyDictionary<string, object?>? componentParameters)
+        IReadOnlyDictionary<string, object?>? componentParameters,
+        bool preventStreamingRendering)
     {
         var endpointHtmlRenderer = httpContext.RequestServices.GetRequiredService<EndpointHtmlRenderer>();
         return endpointHtmlRenderer.Dispatcher.InvokeAsync(async () =>
@@ -43,45 +43,44 @@ internal static class RazorComponentEndpoint
 
             await using var writer = CreateResponseWriter(httpContext.Response.Body);
 
-            try
+            // Note that we always use Static rendering mode for the top-level output from a RazorComponentResult,
+            // because you never want to serialize the invocation of RazorComponentResultHost. Instead, that host
+            // component takes care of switching into your desired render mode when it produces its own output.
+            var htmlContent = (EndpointHtmlRenderer.PrerenderedComponentHtmlContent)(await endpointHtmlRenderer.PrerenderComponentAsync(
+                httpContext,
+                typeof(RazorComponentEndpointHost),
+                RenderMode.Static,
+                hostParameters,
+                waitForQuiescence: preventStreamingRendering));
+
+            // Importantly, we must not yield this thread (which holds exclusive access to the renderer sync context)
+            // in between the first call to htmlContent.WriteTo and the point where we start listening for subsequent
+            // streaming SSR batches. Otherwise some other code might dispatch to the renderer sync context and cause
+            // a batch that would get missed.
+            htmlContent.WriteTo(writer, HtmlEncoder.Default); // Don't use WriteToAsync, as per the comment above
+
+            if (!htmlContent.QuiescenceTask.IsCompleted)
             {
-                // Note that we always use Static rendering mode for the top-level output from a RazorComponentResult,
-                // because you never want to serialize the invocation of RazorComponentResultHost. Instead, that host
-                // component takes care of switching into your desired render mode when it produces its own output.
-                var htmlContent = (EndpointHtmlRenderer.PrerenderedComponentHtmlContent)(await endpointHtmlRenderer.PrerenderComponentAsync(
-                    httpContext,
-                    typeof(RazorComponentEndpointHost),
-                    RenderMode.Static,
-                    hostParameters,
-                    waitForQuiescence: false));
+                endpointHtmlRenderer.OnContentUpdated(htmlComponents =>
+                    EmitStreamingRenderingUpdate(htmlComponents, writer));
 
-                // Importantly, we must not yield this thread (which holds exclusive access to the renderer sync context)
-                // in between the first call to htmlContent.WriteTo and the point where we start listening for subsequent
-                // streaming SSR batches. Otherwise some other code might dispatch to the renderer sync context and cause
-                // a batch that would get missed.
-                htmlContent.WriteTo(writer, HtmlEncoder.Default); // Don't use WriteToAsync, as per the comment above
-
-                if (!htmlContent.QuiescenceTask.IsCompleted)
+                try
                 {
-                    endpointHtmlRenderer.OnContentUpdated(htmlComponents =>
-                        EmitStreamingRenderingUpdate(htmlComponents, writer));
-
                     await writer.FlushAsync(); // Make sure the initial HTML was sent
                     await htmlContent.QuiescenceTask;
                 }
-            }
-            catch (NavigationException navigationException)
-            {
-                HandleNavigation(httpContext, writer, navigationException.Location);
-            }
-            catch (Exception ex) when (httpContext.Response.HasStarted)
-            {
-                // We already started the response so we have no choice but to return a 200 with HTML and will
-                // have to communicate the error information within that (and leave the client-side JS to display
-                // it). But we also rethrow because we want the regular pipeline to treat this like any other
-                // unhandled exception.
-                HandleException(httpContext, writer, ex);
-                throw;
+                catch (NavigationException navigationException)
+                {
+                    HandleNavigationAfterResponseStarted(writer, navigationException.Location);
+                }
+                catch (Exception ex)
+                {
+                    HandleExceptionAfterResponseStarted(httpContext, writer, ex);
+
+                    // The rest of the pipeline can treat this as a regular unhandled exception
+                    // TODO: Is this really right? I think we'll terminate the response in an invalid way.
+                    throw;
+                }
             }
 
             // Invoke FlushAsync to ensure any buffered content is asynchronously written to the underlying
@@ -93,6 +92,7 @@ internal static class RazorComponentEndpoint
 
     private static void EmitStreamingRenderingUpdate(IEnumerable<HtmlComponent> htmlComponents, TextWriter writer)
     {
+        writer.Write("<blazor-ssr>");
         foreach (var entry in htmlComponents)
         {
             // This relies on the component producing well-formed markup (i.e., it can't have a closing
@@ -103,10 +103,13 @@ internal static class RazorComponentEndpoint
             entry.WriteHtmlTo(writer);
             writer.Write("</template>");
         }
+        writer.Write("</blazor-ssr>");
     }
 
-    private static void HandleException(HttpContext httpContext, TextWriter writer, Exception exception)
+    private static void HandleExceptionAfterResponseStarted(HttpContext httpContext, TextWriter writer, Exception exception)
     {
+        // We already started the response so we have no choice but to return a 200 with HTML and will
+        // have to communicate the error information within that
         var env = httpContext.RequestServices.GetRequiredService<IWebHostEnvironment>();
         var options = httpContext.RequestServices.GetRequiredService<IOptions<RazorComponentsEndpointsOptions>>();
         var showDetailedErrors = env.IsDevelopment() || options.Value.DetailedErrors;
@@ -119,19 +122,11 @@ internal static class RazorComponentEndpoint
         writer.Write("</template>");
     }
 
-    private static void HandleNavigation(HttpContext httpContext, TextWriter writer, string destinationUrl)
+    private static void HandleNavigationAfterResponseStarted(TextWriter writer, string destinationUrl)
     {
-        var response = httpContext.Response;
-        if (response.HasStarted)
-        {
-            writer.Write("<template blazor-type=\"redirection\">");
-            writer.Write(destinationUrl);
-            writer.Write("</template>");
-        }
-        else
-        {
-            response.Redirect(destinationUrl);
-        }
+        writer.Write("<template blazor-type=\"redirection\">");
+        writer.Write(destinationUrl);
+        writer.Write("</template>");
     }
 
     private static TextWriter CreateResponseWriter(Stream bodyStream)
